[뉴스 기사 관리] 서비스 및 컨트롤러 메서드 추가

+++ src/main/java/com/example/monew/domain/article/dto/ArticleDto.java
@@ -9,7 +9,7 @@ public record ArticleDto(
         String title,
         LocalDateTime publishDate,
         String summary,
-        Integer commentCount,
-        Integer viewCount,
+        Long commentCount,
+        Long viewCount,
         Boolean viewedByMe // 요청자의 조회 여부
 ) {}
